Refactor quote schemas a bit
